Refactor DateTime class to Temporal
