+++ src/main/java/carbon/Carbon.java
@@ -3,10 +3,7 @@ package carbon;
 import carbon.exceptions.InvalidArgumentException;
 import carbon.exceptions.InvalidCommandException;
 import carbon.exceptions.InvalidFileFormatException;
-import carbon.task.Deadline;
-import carbon.task.Event;
-import carbon.task.Task;
-import carbon.task.Todo;
+import carbon.task.*;
 
 import java.io.File;
 import java.io.FileNotFoundException;
